pyproject.toml
[project]
name = "pushover-bot"
version = "1.0.0"
description = "Telegram bot for Pushover emergency notifications"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.15.0",
    "apscheduler>=3.10.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0",
    "redis>=5.0.0",
    "pillow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
